viewer/requirements.txt
coloredlogs
eclipse-zenoh
pycdr2
numpy-quaternion
numpy_ringbuffer